VotePlugin: Re-enable voting on milestones, refs #4546, #7592 and #10942.
`delete_vote` obviously needed to not depend on a `req` object to work.
That went unnoticed, because it is called only by change listeners for now.

`set_vote` was not ready for unversioned resources, and a subtle insanity in
`resource_from_path` surfaced during early testing with milestones too.

Method name changes are done for clarity, that a single method call probably
deals with multiple vote entries.

git-svn-id: https://trac-hacks.org/svn/voteplugin/trunk@13093 7322e99d-02ea-0310-aa39-e9a107903beb